+++ backend/models/request_institution.py
@@ -4,7 +4,7 @@
 from . import Invite
 from . import Request
 from google.appengine.ext import ndb
-from util.provider_institutions import get_deciis
+from . import get_deciis
 from custom_exceptions.fieldException import FieldException
 from send_email_hierarchy import AcceptedInstitutionEmailSender
